+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [1.1.0](https://github.com/LyCecilion/ret2cli/compare/v1.0.0...v1.1.0) - 2026-08-03
+## [1.1.0] - 2026-08-03 - Bloom in Two
 
 ### Added